iex/midi/tests: Fix wrong operations data beeing used
MidiOperations data is set based on a the midi file name.
The tests set their default data based on a relative file path, but
ScoreRW::readScore turns relative paths to absolute and imports using
those. This caused two data entries to be created per file and the
set defaults to never be used.